Update to Debian 11 and Node 16
The block is currently based on Debian 10, however its mesa version
is too old for RPi kernel 5.15.9 and up which results in X server
crashing on startup.

This patch updates the base to Debian 11 which seems to work fine.
It also bumps node from 12 to 16 since 12 is quite old at this moment
and not available for all the architectures in combination
with Debian 11.

Change-type: minor
Signed-off-by: Michal Toman <michalt@balena.io>